Learning is Occuring
Found a menu select ability from the internet.  Added it to the
thirdfile.py to be figured out later with different text.  Currently
runs, but does not exit loop.  Have to force quit.  Also, raw_input was
a python 2 way to get keyboard presses.  In python 3 it has been renamed
to just input().